文件加载的处理

+++ PhotonIM/PhotonIM/PhotonIMSource/Components/MenuView/PhotonMenuView.m
@@ -56,7 +56,7 @@
         PhotonChatBaseItem *item = (PhotonChatBaseItem *)self.obj;
         if (item.fromType == PhotonChatMessageFromSelf) {
             if ([[item userInfo] messageType] == PhotonIMMessageTypeText) {
-                if ([item canWithDrawMsg] && [[item userInfo] messageStatus] != PhotonIMMessageStatusFailed) {
+                if ([item canWithDrawMsg] && [[item userInfo] messageStatus] != PhotonIMMessageStatusFailed && [[item userInfo] messageStatus] != PhotonIMMessageStatusSending) {
                      [self.menuController setMenuItems:@[copyItem,transpondItem,reDrawItem,deleteItem]];
                 }else{
                      [self.menuController setMenuItems:@[copyItem,transpondItem,deleteItem]];
